Makefile: always use GOPROXY
USe the offical golang goproxy, the fedora package by default uses the
direct mode which pull from the repos directly. However if the
repository is removed or has other errors that will fail while the
goproxy keeps a cache around.

In this particullar case go.opencensus.io has an expired SSL cert and
thus we fail to download it. This becomes an issue in CI as we need to
pull everything to check that vendor tree is clean.

Podman and buildah seems to already set this so we match them with that
change.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>