pkg/detach: fix broken Copy() detach sequence
The code only could handle the detach sequence when it read one byte at
a time. This obviously is not correct and lead to some issues for my
automated test in my podman PR[1] where I added some automated tests
for detaching and the read part is really undefined and depends on the
input side/kernel scheduling on how much we read at once.

This is large rework to make the code check for the key sequence across
the entire buffer. That is of course more work but it needs to happen
for this to work correctly.

I guess the only reason why this was never noticed is because normally
user detach manually by typing and not in an automated way which is much
slower and thus likely reads the bytes one by one.

I added new test to actually confirm the behavior. And to ensure this
works with various read sizes I made it a fuzz test. I had this running
for a while and did not spot any issues there. The old code fails
already on the simple test cases.

[1] https://github.com/containers/podman/pull/25083

Signed-off-by: Paul Holzinger <pholzing@redhat.com>