Fix getLibpodTmpDir() on Windows
Changes the function to return an absolute path to a temporary
directory. Without this, EngineConfig.Validate() will return an error
on Windows when using the default values, as /run/libpod is not an
absolute directory on Windows.

As C:\run is not a temporary directory on Windows, changes the path to
be within the Temp folder.

Signed-off-by: Sebastian Soto <ssoto@redhat.com>