libnetwork/resolvconf: add new KeepHostSearches/Options
Using one KeepHostServers that controls the override of nameservers,
search domains and options is not good enough.

With netavark 1.15 we dropped the dns.podman search domain[1] as this
always overwrote the host search domains which was not correct. However
that in turn caused a new issue[2] that a container name might now get
resolved to a search domain from the host first.
To fix that we either need to revert the dns.podman change or add the
ndots:0 option in resolv.conf. Whatever we end up doing we will need one
of KeepHostSearches or KeepHostOptions in podman to populate resolv.conf
correctly so that we don't overwrite the host domains/options but still
can overwrite the nameservers as we want to force aardvark-dns only as
nameserver so resolvers cannot bypass it.

[1] https://github.com/containers/netavark/pull/1214
[2] https://github.com/containers/podman/issues/26198

Signed-off-by: Paul Holzinger <pholzing@redhat.com>