ci: add govulncheck job
The main idea here is to find any possible vulnerabilities in
dependencies that would require upgrades.

Note that this setup ignores Go stdlib vulnerabilities by using latest
Go version (which is a default for govulncheck-action), as this is a
library and we should not care much about specific Go version used.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>