libnetwork/etchosts: add host.containers.internal for machine
When running inside podman machine we cannot use the normal ip lookup
logic for host.containers.internal because that should refer to the
actual host system and not the VM ip.

gvproxy already resolves the host.containers.internal name correctly but
the issue is when a users wants to set a custom name via --add-host
foobar:host-gateway then we need to know the actual ip to replace the
host-gateway part. Right now we just always error which is not good.

To fix this just look up the name ourselves so we can add it to
/etc/hosts.

Fixes: containers/podman#21681

Signed-off-by: Paul Holzinger <pholzing@redhat.com>